fix: rename unstable css
