+++ .storybook/theme-decorator.tsx
@@ -4,9 +4,9 @@ import { light, dark } from '@charcoal-ui/theme'
 import { useDarkMode } from './use-dark-mode'
 import { TokenInjector, themeSelector, themeSetter } from '@charcoal-ui/styled'
 
-import '../packages/theme/src/css/_token_v1.css'
-import '../packages/theme/src/css/_variables_dark.css'
-import '../packages/theme/src/css/_variables_light.css'
+import '../packages/theme/src/css/v1/remap.css'
+import '../packages/theme/src/css/v2/dark.css'
+import '../packages/theme/src/css/v2/light.css'
 import '../packages/react/dist/index.css'
 
 const setter = themeSetter()
