+++ packages/theme/src/css/v2/dark.css
