+++ packages/theme/src/css/v2/light.css.d.ts
