--- packages/theme/src/docs/TokenV2.mdx
@@ -1,168 +1,0 @@
-import { Meta } from '@storybook/addon-docs/blocks'
-
-<Meta title="theme/Design Token 2.0" />
-
-# Design Token 2.0 セットアップガイド
-
-## インストール
-
-```bash
-npm install --save-dev @charcoal-ui/theme @charcoal-ui/tailwind-config
-npm install @charcoal-ui/react
-```
-
-```tsx
-// Design Token 1.0互換モード
-import '@charcoal-ui/theme/css/_token_v1.css'
-
-// Design Token 2.0
-import '@charcoal-ui/theme/css/_variables_dark.css'
-import '@charcoal-ui/theme/css/_variables_light.css'
-
-// コンポーネント実装のCSS
-import '@charcoal-ui/react/dist/index.css'
-```
-
-## 概要
-
-Design Token 2.0 は `@charcoal-ui/react` のコンポーネントで使われる CSS Variables を提供します。
-
-Token V2 の変数は `.ch-token-v2` が付いた要素の内側で有効になります。既存の Design Token 1.0 と同じ画面で共存できるように、まず CSS を読み込み、Token V2 を使いたい範囲だけに `.ch-token-v2` を付けてください。
-
-## CSS を読み込む
-
-アプリケーションのエントリポイントで、Token V1 互換 CSS、Token V2 の CSS、React コンポーネントの CSS を読み込みます。
-
-```tsx
-import '@charcoal-ui/theme/css/_token_v1.css'
-import '@charcoal-ui/theme/css/_variables_dark.css'
-import '@charcoal-ui/theme/css/_variables_light.css'
-import '@charcoal-ui/react/dist/index.css'
-```
-
-CSS を読み込むだけでは、ライトテーマの Token V2 は有効になりません。`@charcoal-ui/react` のコンポーネントを Token V2 で表示したい範囲に `.ch-token-v2` を付けます。
-
-## コンポーネントで Token V2 を有効にする
-
-もっとも小さく導入する場合は、Token V2 で表示したいコンポーネントを `.ch-token-v2` で囲みます。
-
-```tsx
-import { Button, TextField } from '@charcoal-ui/react'
-
-export const FormActions = () => (
-  <div className="ch-token-v2">
-    <TextField label="タイトル" />
-    <Button variant="Primary">保存</Button>
-  </div>
-)
-```
-
-`.ch-token-v2` の内側では、`@charcoal-ui/react` のコンポーネント CSS が参照する `--charcoal-color-*` や `--charcoal-space-*` などの CSS Variables が Token V2 の値になります。`.ch-token-v2` の外側にあるコンポーネントは、従来どおり Token V1 互換の値を参照します。
-
-### 画面全体で有効にする
-
-画面全体を Token V2 に切り替える場合は、アプリケーションのルート要素に `.ch-token-v2` を付けます。
-
-```tsx
-import { Button } from '@charcoal-ui/react'
-
-export const App = () => (
-  <main className="ch-token-v2">
-    <Button variant="Primary">投稿する</Button>
-  </main>
-)
-```
-
-HTML ルートに付けても有効です。
-
-```html
-<html class="ch-token-v2" data-theme="light">
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
-```
-
-### 一部のコンポーネントだけで試す
-
-既存画面へ段階的に導入する場合は、検証したいコンポーネント単位で `.ch-token-v2` を付けます。
-
-```tsx
-import { Button } from '@charcoal-ui/react'
-
-export const CompareButtons = () => (
-  <>
-    <Button>Token V1</Button>
-
-    <span className="ch-token-v2">
-      <Button>Token V2</Button>
-    </span>
-  </>
-)
-```
-
-Token V2 は CSS Variables を継承して効くため、`.ch-token-v2` は対象コンポーネント自身か、その親要素に付けてください。兄弟要素や離れた要素に付けても影響しません。
-
-## テーマの切り替え
-
-テーマは `document.documentElement` の `data-theme` 属性で切り替えます。
-
-```tsx
-document.documentElement.dataset.theme = 'light'
-document.documentElement.dataset.theme = 'dark'
-```
-
-Token V2 の適用範囲は `.ch-token-v2` で指定します。light または `data-theme` を省略した状態では、特定のコンポーネントだけを Token V2 にできます。
-
-```tsx
-export const TokenV2Area = ({ children }: { children: React.ReactNode }) => (
-  <div className="ch-token-v2">{children}</div>
-)
-```
-
-`data-theme` を省略した場合は light として扱われます。dark theme は `:root[data-theme='dark']` に対して変数が定義されるため、HTML ルートで切り替えます。
-
-```html
-<html class="ch-token-v2" data-theme="dark">
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
-```
-
-## Storybook で確認する
-
-Storybook の story で Token V2 を確認する場合は、`parameters.tokenVersion` に `v2` を指定します。
-
-```tsx
-import { Meta, StoryObj } from '@storybook/react-vite'
-import Button from './Button'
-
-export default {
-  title: 'react/Button',
-  component: Button,
-} as Meta<typeof Button>
-
-export const TokenV2: StoryObj<typeof Button> = {
-  parameters: {
-    tokenVersion: 'v2',
-  },
-  render: () => <Button variant="Primary">Button</Button>,
-}
-```
-
-このリポジトリの Storybook decorator では、`parameters.tokenVersion === 'v2'` の story を表示するときに `document.documentElement.classList` へ `.ch-token-v2` を付けています。
-
-```tsx
-document.documentElement.classList.toggle('ch-token-v2', isTokenV2)
-```
-
-そのため、story 側で毎回 wrapper を追加しなくても Token V2 の見た目を確認できます。
-
-## 導入時の注意
-
-`.ch-token-v2` を付け忘れると、React コンポーネントの CSS は Token V2 の値を参照できません。コンポーネントの見た目が想定と違う場合は、まず対象コンポーネント自身か親要素に `.ch-token-v2` が付いているかを確認してください。
-
-Token V1 と Token V2 を同じ画面で比較する場合は、Token V2 にしたい範囲だけを `.ch-token-v2` で囲みます。画面全体へ一度に付けると、内側のすべての `@charcoal-ui/react` コンポーネントが Token V2 の値を参照します。
-
-`@charcoal-ui/theme/unstable-css/*` は Token V2 の変数をグローバルに定義する用途です。既存コンポーネントと段階的に共存させる場合は、`.ch-token-v2` でスコープできる `@charcoal-ui/theme/css/*` を使ってください。
