+++ packages/theme/src/token-object/helpers/is-empty-array.ts
