--- packages/theme/src/unstable-css/_variables_dark.css.d.ts
@@ -1,1 +1,0 @@
-declare module "*.css" { }
